app/requirements.txt
djangorestframework==3.12.4
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-ckeditor==6.1.0
Pillow==8.2.0
djangorestframework-simplejwt==4.7.1
django-rest-auth==0.9.5
django-filter==2.4.0
requests-oauthlib==1.3.0
beautifulsoup4==4.9.3
drf-yasg==1.20.0